Add support for F# and Visual Basic projects.
