+++ src/ripple.Testing/MSBuild/framework_name_detection_harness.cs
@@ -13,7 +13,7 @@ namespace ripple.Testing.MSBuild
             get
             {
                 var element = XElement.Parse(theXml);
-                var project = CsProjFile.For(element);
+                var project = ProjFile.For(element);
 
                 return FrameworkNameDetector.Detect(project);
             }
