+++ src/ripple/ripple.csproj
@@ -161,7 +161,7 @@
     <Compile Include="Model\ValidationResult.cs" />
     <Compile Include="Model\VersionConstraint.cs" />
     <Compile Include="Model\VersionToken.cs" />
-    <Compile Include="MSBuild\CsProjFile.cs" />
+    <Compile Include="MSBuild\ProjFile.cs" />
     <Compile Include="MSBuild\Reference.cs" />
     <Compile Include="MSBuild\ReferenceAttacher.cs" />
     <Compile Include="MSBuild\ReferenceStatus.cs" />
